+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+- Add modelform-uses-unicode check to flag dangerous use of the exclude
+  attribute in ModelForm.Meta.
+
 Version 0.11.1 (25 May 2018), the DjangoCon Heidelberg edittion
 ---------------------------------------------------------------